some classes corrections to meet the corresponding interfaces

+++ src/main/java/org/launchcode/vendormangedinventory/models/daos/Product_WarehouseDao.java
@@ -8,5 +8,5 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Repository
 @Transactional
-public interface Product_WarehouseDao extends CrudRepository<Product_Warehouse, Integer> {
+public interface Product_WarehouseDao extends CrudRepository<Product_Warehouse, Long> {
 }
